mgw: Move endpoint/conn updating logic to helper function
Move code logic doing actions based on endpoint state (and its conns)
out of the CRCX/MDCX MGCP handlers.
Most of that logic present in CRCX and MDCX is actually the same and can
be merged.
This way we can also make the updating logic more complex without
extending the CRCX/MDCX handlers for and more adding spaghetti code.

Change-Id: I24dd15ec76bd6e949a178a3b998b76a44ddbab50