mgcp_client: add support for sending XSIG to MGW
This patch is currently EXPERIMENTAL, not for merging yet.  The
transcoding MGW that is currently planned for ThemWi MSC will
include generation of in-band DTMF in PCMoIP output, and there is
a need to pass GSM DTMF START and DTMF STOP commands from to-be-
extended osmo-msc to this new MGW.  This experimental patch to
libosmo-mgcp-client adds support for the currently envisioned
XSIG command.

Change-Id: I6411898087bd233445c94a55232c3b74f9b472d6