mgcp_client: add support for driving ThemWi transcoder MGW
This patch is currently EXPERIMENTAL, not for merging yet.  There is
an experimental project underway at Themyscira Wireless whereby a more
complete MSC is to be constructed from suitably modified osmo-msc plus
additional components.  In this architecture the modified osmo-msc
will drive a custom, to-be-developed MGCP MGW that will be highly
asymmetric, in contrast with classic osmo-mgw.  This asymmetric MGW
will have an Ater-IP side (Ater-IP is our term for AoIP compressed
speech interface modified with ThemWi RTP payload formats, providing
functionality of classic TDM Ater over IP) and a PCMoIP side, where
the latter is functionally equivalent to 3GPP-specified AoIP with
_uncompressed_ PCM speech (G.711 or CSData).  The Ater-IP side of
this osmo-msc-driven MGW will connect to the BSS, while the PCMoIP
side connects to further MSC components beyond osmo-msc front end.

This experimental patch to libosmo-mgcp-client adds support for the
currently envisioned MGCP dialect to be spoken by this new MGW.

Change-Id: I713147d6567fb30de09b2272478d717093d6170e